Fix RMI-based single-instance for DISPLAY variables with a slash
For security reasons, we switched the technology to communicate
between Fiji instances (which is used to make sure that there is only
one instance) from socket communication to RMI.

Furthermore, to guarantee that Fiji instances running on different
accounts or screens do not get mixed up, we include not only the user
name but also the content of the DISPLAY environment variable in the
name of the file used for the communication.

Since file names cannot contain slashes, we have to take care of
slashes in the DISPLAY variable explicitly.

Noticed by Wayne Rasband.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>