Add a primitive Javassist-based performance profiler
This profiler is loosely modeled on the ideas of OProfile, but uses
Javassist to instrument the loaded classes.

The gist: every instrumented method is added to a list, and gets two
static fields in the same class: the call count and the total amount
of time spent in the method (or in methods called from there).

Unfortunately, it seems that the precision with which Java can measure
CPU time per thread is a bit crude: in my tests, the granularity is
five milliseconds.

To use it, call fiji.PerformanceProfiler as main class with the "real
main class" and its arguments as parameters.

But it is a start at something we can use, re-use and enhance.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>