Fix handling of NoClassDefFoundErrors again
The original fix in 'Modify ImageJ 1.x to avoid overzealous handling
of NoClassDefFoundErrors' was not very well tested by yours truly:
it missed out on the rather important fact that the exception's
message is *not* the class name that was not found in every case: if
a .class file was found in the expected spot which describes another
class (e.g. in a different package), the exception's message reads:
package/Class (wrong name: XYZ).

This was reported as Fiji bug 576 by Jan Eglinger.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>