Move the prefsDir handling to ij-legacy
In Fiji, we store the ImageJ 1.x preferences in the home directory,
not what happens to be the current directory -- this is almost
never what the user wants (in particular when Fiji is installed in
a non-writable location and when Fiji is used by multiple users).

This logic is not specific to life sciences, so off it goes to ImageJ2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>