Special-case MacOSX in Edit>Memory & Threads... again
We tried very hard to reduce the changes required for Fiji to work
properly. Alas, certain parts of ImageJ 1.x make that hard, and in the
case of Memory & Threads..., it is even impossible because there are so
many assumptions baked into ij.plugin.Memory that they cannot be
overridden except by rewriting the complete plugin from scratch. That is
what we used to do, and that is what we now have to reinstate
reluctantly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>