Automatically create a modern ImageJ context upon startup
Earlier, during the development of the 'ij-legacy' branch, we tried
to avoid having to start up an ImageJ context because ImageJ2 was
not yet optimized for start-up time.

However, in the meantime, Curtis Rueden worked quite a bit on this
and it is very quick already (about one second in this developer's
very unscientific test).

So let's just go for it and always make one context available. To
make it smoother, let's do it in a thread so that the GUI shows up
quickly for impatient users.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>