Fix ./fiji --run 'Switch to Modern Mode'
First, we must take care not to create *two* contexts. Fortunately,
this is easy because the way ij-legacy suggests to obtain a context
actually creates one if necessary, and does so in a synchronized
manner.

Second, we must take care to do this on the Event Dispatch Thread if
we want to avoid dead locks because the switcher plugin must run on the
Event Dispatch Thread due to its many interactions with Swing and AWT.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>