Future-proof the IJ1Patcher
We will re-structure ij-legacy in a rather major way soon. The
LegacyExtensions class will lose its setters and configurations will be
performed exclusively by implementing SciJava-common plugins of certain
types (such as LegacyEditor, LegacyAppConfiguration, etc).

This is step one in the transition: use Javassist to access the current
methods to configure the patching.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>