Add a safety net when registering SwitchToModernMode
We just switched to using the SwitchToModernMode class that lives in the
new package name, net.imagej.legacy.

However, we simply cannot rely on the fact that update sites will all be
up-to-date with the bleeding edge all the time. So we *must* deal nicely
with version skew.

So let's fall back to calling SwitchToModernMode in the old package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>