Add a fall-back for the two legacy plugins in case of version skew
When ij-legacy.jar is on the class path and imagej-legacy.jar (which
obsoletes ij-legacy.jar) is not, we still want Fiji to start up
properly, including window title and all.

So let's add back the plugin types with the old package names so that
SciJava Common's annotation indexer picks them up, too, and everything
works whether ij-legacy.jar is available in the class path, or
imagej-legacy.jar, or both.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>