IJ1Patcher: do not rely on imagej-legacy.jar being in the class path
The imagej-legacy classes are not necessarily in the same class path
as the LegacyEnvironment at runtime patching time.

Therefore, the previous fix to detect the presence of the imagej-legacy
classes (which was not tested properly before pushing to master
prematurely) did not work as expected.

Instead, let's detect the presence of the respective .jar files in
the jars/ subdirectory of the location specified by the system property
'ij.dir'.

If the system property 'ij.dir' is not specified at all, let's assume
that we're not even running in an ImageJ.app/ or Fiji.app/, and as
a consequence let's assume that the user has no version skew at all.
This is the case e.g. when running the integration tests of fiji-compat
itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>