Avoid depending on imagej-legacy when falling back on ij-legacy
With the switch from ij-legacy to imagej-legacy, we need to use
interfaces that now live in a different package. The previous attempt
to re-instate support for the former package name, however, *still*
depended on the new package name: by letting OldLegacyEditor extend
FijiLegacyEditor (that implements the interface in the new package
name), we sadly guaranteed that the OldLegacyEditor class could not be
loaded at all in the intended situation (i.e. when the interfaces in
the new package are unavailable).

Hence, let's bite the bullet and just copy code into the already
deprecated classes, and have fall backs that actually fall back in
the case where ij-legacy is available, but imagej-legacy is not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>