DefaultFijiService: initialize as late as possible
What we really want is to initialize *after* the LegacyService has
initialized. The normal way to accomplish this would be to add an
@Parameter on the LegacyService, which would force that service to be
initialized first. Unfortunately, by doing that, we force the class
loader to load the LegacyService prematurely, hosing everything.

So the hacky solution we employ here is instead to listen for when
context creation is totally complete (signified by the
ServicesLoadedEvent being fired), and *then* finally initialize the
DefaultFijiService. We do it this way, rather than a lazy initialization
scheme, because we *do* want the DefaultFijiService to initialize right
away rather than only in response to some API call later.