Be much more stringent when we allow ImageJ 1.x to DWIM
The static method in the Opener class intended to help with converting
grayscale images that have been read as RGB by mistake (or by design:
the GIF actually does not specify any explicit grayscale mode).

It does not work properly, though, because it does not safeguard
e.g. against trying to convert 16-bit grayscale to 8-bit by mistake,
as the name would suggest.

While at it, reset the display range on opened 16-bit sample images,
such as the M51 Galaxy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>